Use semaphore to wait for the initlization of memory hook functions

+++ AddressMap.h
@@ -85,7 +85,7 @@ public:
     {
         int index;
         if (hasAddr(addr, 0, _size - 1, &index)) {
-            for (int i = index; i < _size - 1; i++)
+            for (size_t i = index; i < _size - 1; i++)
                 _map[i] = _map[i+1];
             _size--;
         }
